move MailQueueListener to Subscriber namespace
used only internally; safe to move
